Refactor homepage into multiple page
